# laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentació de la nostra API",
        "description": "",
        "contact": {
            "name": "Equip de Spottunes",
            "email": "a20pedgarguz@inspedralbes.cat"
        },
        "license": {
            "name": "Licencia API",
            "url": "http://www.example.com/licenses/api"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/users/follow/{userId}": {
            "post": {
                "tags": [
                    "Seguidors"
                ],
                "summary": "Segueix a un usuari",
                "description": "Permet a un usuari seguir a un altre usuari per la seva ID.",
                "operationId": "followUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID de l'usuari a seguir",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ara estàs seguint a aquest usuari",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ara estàs seguint a aquest usuari"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ja estàs seguint a aquest usuari",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ja estàs seguint a aquest usuari"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuari no trobat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuari no trobat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autoritzat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token no proporcionat als encapçalaments."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users/unfollow/{userId}": {
            "delete": {
                "tags": [
                    "Seguidors"
                ],
                "summary": "Deixar de seguir a un usuari",
                "description": "Permet a un usuari deixar de seguir a un altre usuari per la seva ID.",
                "operationId": "unfollowUser",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID de l'usuari a deixar de seguir",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ja no estàs seguint a aquest usuari",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Ja no estàs seguint a aquest usuari"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuari no trobat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuari no trobat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autoritzat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token no proporcionat als encapçalaments."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users/followers/{userId}": {
            "get": {
                "tags": [
                    "Seguidors"
                ],
                "summary": "Obtenir seguidors d'un usuari",
                "description": "Obté la llista de seguidors d'un usuari per la seva ID, incloent el compte total de seguidors.",
                "operationId": "getUserFollowers",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID de l'usuari del qual es volen obtenir els seguidors",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Llista de seguidors obtinguda amb èxit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "followers": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "nickname": {
                                                        "type": "string",
                                                        "example": "seguidor123"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "imatges/imatge_seguidor.png"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuari no trobat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuari no trobat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autoritzat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token no proporcionat als encapçalaments."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users/followed/{userId}": {
            "get": {
                "tags": [
                    "Seguidors"
                ],
                "summary": "Obtenir usuaris seguits per un usuari",
                "description": "Obté la llista d'usuaris seguits per un usuari per la seva ID, incloent el compte total d'usuaris seguits.",
                "operationId": "getUserFollowed",
                "parameters": [
                    {
                        "name": "userId",
                        "in": "path",
                        "description": "ID de l'usuari del qual es volen obtenir els usuaris seguits",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Llista d'usuaris seguits obtinguda amb èxit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "count": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "followed": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 2
                                                    },
                                                    "nickname": {
                                                        "type": "string",
                                                        "example": "usuariSeguit123"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "imatges/imatge_usuari_seguidor.png"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Usuari no trobat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Usuari no trobat"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autoritzat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token no proporcionat als encapçalaments."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/users/followers": {
            "get": {
                "tags": [
                    "Seguidors"
                ],
                "summary": "Obtenir seguidors de l'usuari autenticat",
                "description": "Obté la llista d'IDs dels usuaris que segueixen a l'usuari autenticat.",
                "operationId": "getFollowers",
                "responses": {
                    "200": {
                        "description": "Llista de seguidors obtinguda amb èxit",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autoritzat",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Token no proporcionat als encapçalaments."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}